requirements.txt
aiogram==3.25.0
pydantic>=2.12,<2.13
python-dotenv==1.0.1
httpx==0.27.0
